requirements.txt
asyncio
discord.py